Mod name change! Mod is now known as Chests+
