--- src/main/java/dmillerw/cchests/core/proxy/CommonProxy.java
@@ -1,12 +1,0 @@
-package dmillerw.cchests.core.proxy;
-
-import cpw.mods.fml.common.event.FMLPostInitializationEvent;
-import cpw.mods.fml.common.event.FMLPreInitializationEvent;
-
-public class CommonProxy {
-
-	public void preInit(FMLPreInitializationEvent event) {}
-	
-	public void postInit(FMLPostInitializationEvent event) {}
-	
-}
