+++ src/main/java/dmillerw/chestsplus/core/proxy/CommonProxy.java
@@ -1,0 +1,14 @@
+package dmillerw.chestsplus.core.proxy;
+
+import cpw.mods.fml.common.event.FMLPostInitializationEvent;
+import cpw.mods.fml.common.event.FMLPreInitializationEvent;
+
+public class CommonProxy {
+
+	public void preInit(FMLPreInitializationEvent event) {
+	}
+
+	public void postInit(FMLPostInitializationEvent event) {
+	}
+
+}
