+++ src/main/resources/assets/chestsplus/textures/normal_trans.png
